api/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
aspose-words==24.8.0